pyproject.toml
[project]
name = "autotest"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "allure-pytest>=2.15.3",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "selenium>=4.40.0",
]
